arrow_select::coalesce::generic

Struct GenericInProgressArray

pub(crate) struct GenericInProgressArray {
    source: Option<ArrayRef>,
    buffered_arrays: Vec<ArrayRef>,
}

Generic implementation for InProgressArray that works with any type of array.

Internally, this buffers arrays and then calls other kernels such as concat to produce the final array.

Fields
source: Option<ArrayRef>

The current source

buffered_arrays: Vec<ArrayRef>

The buffered array slices

Implementations

impl GenericInProgressArray

pub(crate) fn new() -> Self

Create a new GenericInProgressArray