arrow_select
Module dictionary

Structs

Interner 🔒

A best effort interner that maintains a fixed number of buckets and interns keys based on their hash value

MergedDictionaries

Functions

bytes_ptr_eq 🔒

Performs a cheap, pointer-based comparison of two byte array

compute_values_mask 🔒

Return a mask identifying the values that are referenced by keys in dictionary at the positions indicated by selection

get_masked_values 🔒

Return a Vec containing for each set index in mask, the index and byte value of that index

masked_bytes 🔒

Compute get_masked_values for a [GenericByteArray]

merge_dictionary_values

Given an array of dictionaries and an optional key mask compute a values array containing referenced values, along with mappings from the [DictionaryArray] keys to the new keys within this values array. Best-effort will be made to ensure that the dictionary values are unique

should_merge_dictionary_values

A weak heuristic of whether to merge dictionary values that aims to only perform the expensive merge computation when it is likely to yield at least some return over the naive approach used by MutableArrayData

Type Aliases

PtrEq 🔒

A type-erased function that compares two array for pointer equality