Module arrow_select::filter

Defines filter kernels

Structs

FilterBuilder

A builder to construct FilterPredicate

FilterBytes 🔒

FilterBytes is created from a source [GenericByteArray] and can be used to build a new [GenericByteArray] by copying values from the source

FilterPredicate

A filtering predicate that can be applied to an [Array]

IndexIterator 🔒

An iterator of usize whose index in [BooleanArray] is true

SlicesIterator

An iterator of (usize, usize) each representing an interval [start, end) whose slots of a bitmap [Buffer] are true. Each interval corresponds to a contiguous region of memory to be “taken” from an array to be filtered.

Enums

IterationStrategy 🔒

The iteration strategy used to evaluate FilterPredicate

Constants

FILTER_SLICES_SELECTIVITY_THRESHOLD 🔒

If the filter selects more than this fraction of rows, use SlicesIterator to copy ranges of values. Otherwise iterate over individual rows using IndexIterator

Functions

build_filterDeprecated

Returns a prepared function optimized to filter multiple arrays. Creating this function requires time, but using it is faster than filter when the same filter needs to be applied to multiple arrays (e.g. a multi-column RecordBatch). WARNING: the nulls of filter are ignored and the value on its slot is considered. Therefore, it is considered undefined behavior to pass filter with null values.

filter

Filters an [Array], returning elements matching the filter (i.e. where the values are true).

filter_array 🔒

filter_bits 🔒

Filter the packed bitmask buffer, with predicate starting at bit offset offset

filter_boolean 🔒

filter implementation for boolean buffers

filter_byte_view 🔒

filter implementation for byte view arrays.

filter_bytes 🔒

filter implementation for byte arrays

filter_count 🔒

Counts the number of set bits in filter

filter_dict 🔒

filter implementation for dictionaries

filter_fixed_size_binary 🔒

filter_native 🔒

filter_null_mask 🔒

Computes a new null mask for data based on predicate

filter_primitive 🔒

filter implementation for primitive arrays

filter_record_batch

Returns a new [RecordBatch] with arrays containing only values matching the filter.

filter_run_end_array 🔒

Filter any supported [RunArray] based on a FilterPredicate

prep_null_mask_filter

Remove null values by do a bitmask AND operation with null bits and the boolean bits.

Type Aliases

FilterDeprecated

Function that can filter arbitrary arrays