arrow_select::filter

Function filter_record_batch

pub fn filter_record_batch(
    record_batch: &RecordBatch,
    predicate: &BooleanArray,
) -> Result<RecordBatch, ArrowError>

Returns a filtered [RecordBatch] where the corresponding elements of predicate are true.

This is the equivalent of calling filter on each column of the [RecordBatch].

If multiple record batches (or arrays) need to be filtered using the same predicate array, consider using FilterBuilder to create a single FilterPredicate and then calling FilterPredicate::filter_record_batch. In contrast to this function, it is then the responsibility of the caller to use FilterBuilder::optimize if appropriate.