arrow_select::filter

Struct FilterBytes

struct FilterBytes<'a, OffsetSize> {
    src_offsets: &'a [OffsetSize],
    src_values: &'a [u8],
    dst_offsets: Vec<OffsetSize>,
    dst_values: Vec<u8>,
    cur_offset: OffsetSize,
}

FilterBytes is created from a source [GenericByteArray] and can be used to build a new [GenericByteArray] by copying values from the source

TODO(raphael): Could this be used for the take kernel as well?

Fields
src_offsets: &'a [OffsetSize]src_values: &'a [u8]dst_offsets: Vec<OffsetSize>dst_values: Vec<u8>cur_offset: OffsetSize
Implementations

impl<'a, OffsetSize> FilterBytes<'a, OffsetSize>
where OffsetSize: OffsetSizeTrait,

fn new<T>(capacity: usize, array: &'a GenericByteArray<T>) -> Self
where T: ByteArrayType<Offset = OffsetSize>,

fn get_value_offset(&self, idx: usize) -> usize

Returns the byte offset at idx

fn get_value_range(&self, idx: usize) -> (usize, usize, OffsetSize)

Returns the start and end of the value at index idx along with its length

fn extend_idx(&mut self, iter: impl Iterator<Item = usize>)

Extends the in-progress array by the indexes in the provided iterator

fn extend_slices(&mut self, iter: impl Iterator<Item = (usize, usize)>)

Extends the in-progress array by the ranges in the provided iterator