arrow_select

Module zip

zip: Combine values from two arrays based on boolean mask

Structs

ByteViewScalarImpl 🔒

BytesScalarImpl 🔒

FallbackImpl 🔒

PrimitiveScalarImpl 🔒

ScalarZipper

Zipper for 2 scalars

Statics

EMPTY_ARC 🔒

Traits

ZipImpl 🔒

Impl for creating output array based on a mask

Functions

empty_arc_buffers 🔒

maybe_prep_null_mask_filter 🔒

zip

Zip two arrays by some boolean mask.

zip_impl 🔒