arrow_select::coalesce

Trait InProgressArray

trait InProgressArray:
    Debug
    + Send
    + Sync {
    // Required methods
    fn set_source(&mut self, source: Option<ArrayRef>);
    fn copy_rows(&mut self, offset: usize, len: usize) -> Result<(), ArrowError>;
    fn finish(&mut self) -> Result<ArrayRef, ArrowError>;
}

Incrementally builds up arrays

GenericInProgressArray is the default implementation that buffers arrays and uses other kernels concatenates them when finished.

Some types have specialized implementations for this array types (e.g., StringViewArray, etc.).

Required Methods

fn set_source(&mut self, source: Option<ArrayRef>)

Set the source array.

Calls to Self::copy_rows will copy rows from this array into the current in-progress array

fn copy_rows(&mut self, offset: usize, len: usize) -> Result<(), ArrowError>

Copy rows from the current source array into the in-progress array

The source array is set by Self::set_source.

Return an error if the source array is not set

fn finish(&mut self) -> Result<ArrayRef, ArrowError>

Finish the currently in-progress array and return it as an ArrayRef