arrow_select

Module filter

Defines filter kernels

Structs

FilterBuilder

A builder to construct FilterPredicate

FilterBytes 🔒

FilterBytes is created from a source [GenericByteArray] and can be used to build a new [GenericByteArray] by copying values from the source

FilterPredicate

A filtering predicate that can be applied to an [Array]

IndexIterator 🔒

An iterator of usize whose index in [BooleanArray] is true

SlicesIterator

An iterator of (usize, usize) each representing an interval [start, end) whose slots of a bitmap [Buffer] are true.

Enums

IterationStrategy 🔒

The iteration strategy used to evaluate FilterPredicate

Constants

FILTER_SLICES_SELECTIVITY_THRESHOLD 🔒

If the filter selects more than this fraction of rows, use SlicesIterator to copy ranges of values. Otherwise iterate over individual rows using IndexIterator

Functions

build_filterDeprecated

Returns a prepared function optimized to filter multiple arrays.

filter

Returns a filtered values [Array] where the corresponding elements of predicate are true.

filter_array 🔒

filter_bits 🔒

Filter the packed bitmask buffer, with predicate starting at bit offset offset

filter_boolean 🔒

filter implementation for boolean buffers

filter_byte_view 🔒

filter implementation for byte view arrays.

filter_bytes 🔒

filter implementation for byte arrays

filter_count 🔒

Counts the number of set bits in filter

filter_dict 🔒

filter implementation for dictionaries

filter_fixed_size_binary 🔒

filter_native 🔒

filter_null_mask 🔒

Computes a new null mask for data based on predicate

filter_primitive 🔒

filter implementation for primitive arrays

filter_record_batch

Returns a filtered [RecordBatch] where the corresponding elements of predicate are true.

filter_run_end_array 🔒

Filter any supported [RunArray] based on a FilterPredicate

filter_sparse_union 🔒

filter implementation for sparse unions

filter_struct 🔒

filter implementation for structs

multiple_arrays 🔒

prep_null_mask_filter

Remove null values by do a bitmask AND operation with null bits and the boolean bits.

Type Aliases

FilterDeprecated

Function that can filter arbitrary arrays